# Rename and rework File::Stat as File::Info