--- a/src/crystal/system/file_info.cr
+++ b/src/crystal/system/file_info.cr
@@ -1,0 +1,1 @@
+require "./unix/file_info"