--- a/src/crystal/system/unix/getrandom.cr
+++ b/src/crystal/system/unix/getrandom.cr
@@ -15,7 +15,7 @@ module Crystal::System::Random
       @@getrandom_available = true
     else
       urandom = ::File.open("/dev/urandom", "r")
-      return unless urandom.stat.chardev?
+      return unless urandom.info.type.character_device?
 
       urandom.close_on_exec = true
       urandom.sync = true # don't buffer bytes