… spec/std/file_spec.cr
… src/file/stat.cr