I've aimed to make `File::Info` easier to use, more portable, and nicer than `File::Stat`.