@RX14 2018-01-14
@asterite yeah, while debugging I found that I really wanted to be able to see the permissions better so I implemented that.