@RX14 2018-01-14
@ysbaddaden Information has been intentionally left out to make it portable. `atime` is disabled on most modern distros, either with `relatime` or `noatime` mount options (and nobody uses it) and `ctime` is usually not what people want or think it is (they want `btime` which basically no filesystems support). `mtime` is really the only widely supported and actually useful time. Go only provides modification time too.

I'll benchmark old stat vs new stat but i'm not sure that a bit of bit twiddling is going to make much difference when put next to the price of a syscall, and making it wrap `LibC::Stat` would move platform-specifics into the definition of `File::Info` instead of just the construction. Keep in mind that on windows I'm not going to be using `LibC::Stat` at all, but recostructing the same information from `GetFileInformationByHandle`.