@RX14 2018-01-14
I do wonder whether we should rename `File.stat?` as `File.info?` and have `File.info?(follow_symlink: false)`.

Also i'm probably going to just rename `File::Type::SymbolicLink` to `File::Type::Symlink`, or add an alias.