@asterite 2018-01-14
I actually learned in this PR that `lstat` doesn't (or does) follow symlinks, so a named argument to select that behavior is better in my opinion. `lstat` is too cryptic. So maybe it's better to just use `info`.

I believe Ruby started with "works only in unix" and eventually had to keep all those cryptic, unix-specific names. Go for example doesn't do this, it provides a clear interface to the computer without using ancient strange names. I think we should do the same, even if it means a bigger barrier to those coming from Ruby.