@RX14 2018-01-15
@asterite its called stat and lstat on go too...