@asterite 2018-01-15
Maybe because go isn't that flexible with method names.