@RX14 2018-01-15
@asterite I do agree it should be `File.info`. (eventually `Path#info` and `File#info` for stat/lstat and fstat respectively)