@RX14 2018-01-21
Updated this PR to take into account the above suggestions.