@RX14 2018-01-21
Looks like I accidentally made this PR depend on https://github.com/crystal-lang/crystal/pull/5619...