@RX14 2018-01-22
> Why change stat naming to info? it's still an abbreviation and doesn't tell any better.

`info` implies you're getting additional data, or metadata, about something. `stat` (if you can decode what it means at all) implies to me statistics about the fd/filename, which is absolutely not the case.

> Why decode & copy everything (check each file type, decode/copy permissions, generate Time instances) instead of using a wrapper for on-demand generation? Sure File::Stat would be a mere delegation struct for the underlying POSIX vs Windows implementations, mostly to hold documentation, but the optimizer will inline calls, or we can just flag them with @[AlwaysInline].

Because it's shorter code (look at the length of `src/file/info.cr` vs `src/file/stat.cr`) and easier to read and all the platform-specifics are nicely contained in a single method in `src/crystal/system`. And I benchmarked it to show that the performance impact was negligible.

> Having to type .info.type.directory? over .stat.directory? can feel minor, but that's still an unwelcomed hick; forward calls from the struct to the enum, please.

There's no point. `fd.info.type.directory?` can never return true since we don't expose `info` on `opendir` FDs, and every other time you can just use `Dir.exists?` or `File.directory?`. If you're trying to get multiple bits of info out of a single stat call, your code length is hardly going to be impacted hugly by 5 extra characters.

> ctime & atime are still missing (!)

They're not useful, and actively misleading, as I explained before:

> Information has been intentionally left out to make it portable. `atime` is disabled on most modern distros and OSes, either with `relatime` or `noatime` mount options (they're disabled because nobody uses it) and `ctime` is usually not what people want or think it is (they want `btime` which basically no filesystems support). `mtime` is really the only widely supported and actually useful time. Go only provides modification time.