@j8r 2018-01-22
It would be too bad to remove the `ctime` feature: sometimes we want to know when a file has been renamed or moved. Why not having a `File.info(attributes: false)`? Maybe a `access: false` for `atime`, not sure.