@asterite 2018-01-23
Just to add a bit more to this, the names and API were copied from Go. I think "info" is better than "stat" as it represents information, not statistics (well, it could be "status", but I don't know).

I also searched other languages. For example in Elixir it's called [File.Stat](https://hexdocs.pm/elixir/File.Stat.html) but right in the beginning of those docs it says "A struct that holds file **information**. In Erlang, this struct is represented by a **:file_info** record". So why did they chose "Stat" for a name? :-P

Then we have [Julia](https://docs.julialang.org/en/release-0.4/stdlib/file/#Base.stat): it's called "stat", but "contain information about the file".

I don't know about the removal of "ctime" and "atime", but since Go doesn't provide that I guess that's fine (Go developers know more than us).