@RX14 2018-01-23
To be clear - I'm not super invested in removing ctime and atime. I just started from the interface that Go provides and until now nobody's given me a usecase for atime and ctime.