@straight-shoota 2018-01-23
I'd say atime and ctime should not really the issue here. I don't see a usecase for them either, but if one emerges, they can easily be re-added later. The same goes for abbreviations for `file.type`. Let's not bother with these details that don't matter much to the general implementation.

Info instead of stat is fine, I guess. An experienced developer might be more likely looking for "stat" but will get around with info wich is IMHO a better name that's more descriptive and easier to understand without any background.