@lbguilherme 2018-01-23
My argument against `stat` and for `info` is that stat is a POSIX naming, and we are going to support non-POSIX operating systems (Windows), better have a more agnostic stdlib, so it feels native for everybody.