@j8r 2018-01-23
A use case for `ctime`: print the last time the file has been moved – can be very useful to know the real date since a file is in its directory. Even more with multiple mount points, but don't know if the `mtime` of a file changes if moved from one to another (e.g. a NFS).

All we need is to add a parameter that adds the last change time of attributes.