@RX14 2018-01-23
@j8r if we include `atime` and `ctime` we'd include it all the time, not optionally.

My complaint though is that you described what `ctime` is *for* but not an actual usecase. When do you need to know this?