@j8r 2018-01-23
@RX14 For backups, when used with retention time. For example you keep the archive in a place, then move it to a NFS.

As I said above, why not having `File.info(attributes: false)`? When true, it returns `ctime`, false the `mtime`?