@asterite 2018-01-23
@j8r it's not clear what you mean when you say `attributes: false`. It returns an object that suddenly has more attributes? This is a compiled language, that's not possible.